Browse: collapse browse.$lang back into a single route file
The route had a loader + validateSearch + staticData, so it can't be a
component-only createLazyFileRoute. But splitting it into an eager config
half + a .lazy component half was needless ceremony: with autoCodeSplitting
enabled, a plain single-file route (like $lang.feed.tsx / $lang.tsx) already
gets its component split into its own on-demand chunk. Merged the two files
back into one browse.$lang.tsx and deleted browse.$lang.lazy.tsx.

Verified: component still builds to a standalone browse._lang chunk
(~7.5 kB gz) that is not in the initial payload — identical laziness, one
fewer file, consistent with the rest of the route tree.
